переделала

+++ test/Program.cs
@@ -3,9 +3,9 @@
 
 //b1 = 2, k1 = 5, b2 = 4, k2 = 9 -> (-0,5; -0,5)
 
-int[] CrsPnt(int[,] myArray)
+double[] CrsPnt(double[,] myArray)
 {
-    int[] crossPoint = new int[2];
+    double[] crossPoint = new double[2];
     crossPoint[0] = (myArray[1,1] - myArray[0,1]) / (myArray[0,0] - myArray[1,0]);
     crossPoint[1] = crossPoint[0] * myArray[0,0] + myArray[0,1];
     return crossPoint;
@@ -19,7 +19,7 @@ Console.Write("Введите k2: ");
 int k2 = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите b2: ");
 int b2 = Convert.ToInt32(Console.ReadLine());
-int[,] myArray = {{k1, b1}, {k2, b2}};
+double[,] myArray = {{k1, b1}, {k2, b2}};
 Console.WriteLine(String.Join(" ", CrsPnt(myArray)));
